refactor react 19 - vite - useEffect hook
