--- 04-frameworks/01-react/03-react-hooks/03-component-did-onload/.babelrc
@@ -1,7 +1,0 @@
-{
-  "presets": [
-    "@babel/preset-env",
-    "@babel/preset-typescript",
-    "@babel/preset-react"
-  ]
-}
