+++ 04-frameworks/01-react/03-react-hooks/03-component-mount/vite.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from "vite";
+import checker from "vite-plugin-checker";
+import react from "@vitejs/plugin-react";
+
+export default defineConfig({
+  plugins: [checker({ typescript: true }), react()],
+});
