--- 04-frameworks/01-react/03-react-hooks/04-component_unmount/webpack.config.js
@@ -1,54 +1,0 @@
-const HtmlWebpackPlugin = require("html-webpack-plugin");
-const path = require("path");
-const basePath = __dirname;
-
-module.exports = {
-  context: path.join(basePath, "src"),
-  resolve: {
-    extensions: [".js", ".ts", ".tsx"],
-  },
-  entry: {
-    app: ["./index.tsx", "./styles.css"],
-  },
-  devtool: "eval-source-map",
-  stats: "errors-only",
-  output: {
-    filename: "[name].[chunkhash].js",
-  },
-  module: {
-    rules: [
-      {
-        test: /\.tsx?$/,
-        exclude: /node_modules/,
-        loader: "babel-loader",
-      },
-      {
-        test: /\.(png|jpg)$/,
-        type: "asset/resource",
-      },
-      {
-        test: /\.html$/,
-        loader: "html-loader",
-      },
-      {
-        test: /\.css$/,
-        exclude: /node_modules/,
-        use: [
-          {
-            loader: "style-loader",
-          },
-          {
-            loader: "css-loader",
-          },
-        ],
-      },
-    ],
-  },
-  plugins: [
-    //Generate index.html in /dist => https://github.com/ampedandwired/html-webpack-plugin
-    new HtmlWebpackPlugin({
-      filename: "index.html", //Name of file in ./dist/
-      template: "index.html", //Name of template in ./src
-    }),
-  ],
-};
